+++ src/builder/pod/mod.rs
@@ -15,6 +15,8 @@ use k8s_openapi::{
 };
 use std::collections::BTreeMap;
 
+use super::{ListenerOperatorVolumeSourceBuilder, ListenerReference};
+
 /// A builder to build [`Pod`] objects.
 ///
 #[derive(Clone, Default)]
@@ -143,6 +145,138 @@ impl PodBuilder {
 
     pub fn add_volumes(&mut self, volumes: Vec<Volume>) -> &mut Self {
         self.volumes.get_or_insert_with(Vec::new).extend(volumes);
+        self
+    }
+
+    /// Add a [`Volume`] for the storage class `listeners.stackable.tech` with the given listener
+    /// class.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// # use stackable_operator::builder::PodBuilder;
+    /// # use stackable_operator::builder::ContainerBuilder;
+    /// let pod = PodBuilder::new()
+    ///     .metadata_default()
+    ///     .add_container(
+    ///         ContainerBuilder::new("container")
+    ///             .unwrap()
+    ///             .add_volume_mount("listener", "/path/to/volume")
+    ///             .build(),
+    ///     )
+    ///     .add_listener_volume_by_listener_class("listener", "nodeport")
+    ///     .build()
+    ///     .unwrap();
+    ///
+    /// assert_eq!("\
+    /// apiVersion: v1
+    /// kind: Pod
+    /// metadata: {}
+    /// spec:
+    ///   affinity: {}
+    ///   containers:
+    ///   - name: container
+    ///     volumeMounts:
+    ///     - mountPath: /path/to/volume
+    ///       name: listener
+    ///   enableServiceLinks: false
+    ///   volumes:
+    ///   - ephemeral:
+    ///       volumeClaimTemplate:
+    ///         metadata:
+    ///           annotations:
+    ///             listeners.stackable.tech/listener-class: nodeport
+    ///         spec:
+    ///           accessModes:
+    ///           - ReadWriteMany
+    ///           resources:
+    ///             requests:
+    ///               storage: '1'
+    ///           storageClassName: listeners.stackable.tech
+    ///     name: listener
+    /// ", serde_yaml::to_string(&pod).unwrap())
+    /// ```
+    pub fn add_listener_volume_by_listener_class(
+        &mut self,
+        volume_name: &str,
+        listener_class: &str,
+    ) -> &mut Self {
+        self.add_volume(Volume {
+            name: volume_name.into(),
+            ephemeral: Some(
+                ListenerOperatorVolumeSourceBuilder::new(&ListenerReference::ListenerClass(
+                    listener_class.into(),
+                ))
+                .build(),
+            ),
+            ..Volume::default()
+        });
+        self
+    }
+
+    /// Add a [`Volume`] for the storage class `listeners.stackable.tech` with the given listener
+    /// name.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// # use stackable_operator::builder::PodBuilder;
+    /// # use stackable_operator::builder::ContainerBuilder;
+    /// let pod = PodBuilder::new()
+    ///     .metadata_default()
+    ///     .add_container(
+    ///         ContainerBuilder::new("container")
+    ///             .unwrap()
+    ///             .add_volume_mount("listener", "/path/to/volume")
+    ///             .build(),
+    ///     )
+    ///     .add_listener_volume_by_listener_name("listener", "preprovisioned-listener")
+    ///     .build()
+    ///     .unwrap();
+    ///
+    /// assert_eq!("\
+    /// apiVersion: v1
+    /// kind: Pod
+    /// metadata: {}
+    /// spec:
+    ///   affinity: {}
+    ///   containers:
+    ///   - name: container
+    ///     volumeMounts:
+    ///     - mountPath: /path/to/volume
+    ///       name: listener
+    ///   enableServiceLinks: false
+    ///   volumes:
+    ///   - ephemeral:
+    ///       volumeClaimTemplate:
+    ///         metadata:
+    ///           annotations:
+    ///             listeners.stackable.tech/listener-name: preprovisioned-listener
+    ///         spec:
+    ///           accessModes:
+    ///           - ReadWriteMany
+    ///           resources:
+    ///             requests:
+    ///               storage: '1'
+    ///           storageClassName: listeners.stackable.tech
+    ///     name: listener
+    /// ", serde_yaml::to_string(&pod).unwrap())
+    /// ```
+    pub fn add_listener_volume_by_listener_name(
+        &mut self,
+        volume_name: &str,
+        listener_name: &str,
+    ) -> &mut Self {
+        self.add_volume(Volume {
+            name: volume_name.into(),
+            ephemeral: Some(
+                ListenerOperatorVolumeSourceBuilder::new(&ListenerReference::ListenerName(
+                    listener_name.into(),
+                ))
+                .build(),
+            ),
+            ..Volume::default()
+        });
         self
     }